requirements.txt
openai
python-dotenv
pydantic
google-generativeai
Pillow
requests
moviepy
